+++ src/__tests__/methods/fetch/index.test.ts
@@ -1,4 +1,4 @@
-import { tsFetch } from "../../fetch";
+import { tsFetch } from "../../../methods/fetch";
 
 // Mock the fetch API
 global.fetch = jest.fn();
