+++ src/__tests__/methods/patch/index.test.ts
@@ -1,15 +1,15 @@
-import { tsPut } from "../../put";
-import { SimpleResponse } from "../../types";
+import { tsPatch } from "../../../methods/patch";
+import { SimpleResponse } from "../../../types";
 
 // Mocking the global fetch function
 global.fetch = jest.fn();
 
-describe("tsPut", () => {
+describe("tsPatch", () => {
   afterEach(() => {
     jest.clearAllMocks();
   });
 
-  it("should successfully make a PUT request and return the expected data", async () => {
+  it("should successfully make a PATCH request and return the expected data", async () => {
     // Mock response data and headers
     const mockData = { success: true };
     const mockHeaders = new Headers();
@@ -21,11 +21,11 @@ describe("tsPut", () => {
 
     (global.fetch as jest.Mock).mockResolvedValue(mockResponse);
 
-    const url = "https://example.com/put";
+    const url = "https://example.com/patch";
     const requestBody = { name: "John Doe" };
     const requestHeaders = { Authorization: "Bearer token" };
 
-    const result: SimpleResponse<typeof mockData> = await tsPut(
+    const result: SimpleResponse<typeof mockData> = await tsPatch(
       url,
       requestBody,
       requestHeaders,
@@ -33,18 +33,16 @@ describe("tsPut", () => {
 
     // Check that the correct fetch call was made
     expect(fetch).toHaveBeenCalledWith(url, {
-      method: "PUT",
+      method: "PATCH",
       body: requestBody,
       headers: {
         Authorization: "Bearer token",
       },
     });
 
-    // Check the result
     expect(result.data).toEqual(mockData);
     expect(result.status).toBe(200);
-    expect(result.headers.has("Authorization")).toBe(true);
-    expect(result.headers.get("Authorization")).toBe("Bearer token");
+    expect(result.headers).toBeInstanceOf(Headers); // Check if the headers is an instance of Headers
   });
 
   it("should throw an error when the response status is not OK (e.g., 400)", async () => {
@@ -52,12 +50,12 @@ describe("tsPut", () => {
     const mockResponse = new Response(null, { status: 400 });
     (global.fetch as jest.Mock).mockResolvedValue(mockResponse);
 
-    const url = "https://example.com/put";
+    const url = "https://example.com/patch";
     const requestBody = { name: "John Doe" };
     const requestHeaders = {};
 
-    await expect(tsPut(url, requestBody, requestHeaders)).rejects.toThrow(
-      "Network response status 400 with URL: https://example.com/put",
+    await expect(tsPatch(url, requestBody, requestHeaders)).rejects.toThrow(
+      "Network response status 400 with URL: https://example.com/patch",
     );
   });
 
@@ -66,11 +64,11 @@ describe("tsPut", () => {
     const errorMessage = "Network failure";
     (global.fetch as jest.Mock).mockRejectedValue(new Error(errorMessage));
 
-    const url = "https://example.com/put";
+    const url = "https://example.com/patch";
     const requestBody = { name: "John Doe" };
     const requestHeaders = {};
 
-    await expect(tsPut(url, requestBody, requestHeaders)).rejects.toThrow(
+    await expect(tsPatch(url, requestBody, requestHeaders)).rejects.toThrow(
       errorMessage,
     );
   });
@@ -79,11 +77,11 @@ describe("tsPut", () => {
     // Simulate an unexpected error type
     (global.fetch as jest.Mock).mockRejectedValue("Non-error");
 
-    const url = "https://example.com/put";
+    const url = "https://example.com/patch";
     const requestBody = { name: "John Doe" };
     const requestHeaders = {};
 
-    await expect(tsPut(url, requestBody, requestHeaders)).rejects.toThrow(
+    await expect(tsPatch(url, requestBody, requestHeaders)).rejects.toThrow(
       "An unknown error occurred",
     );
   });
