+++ src/methods/delete/index.ts
